Further refactor of touchtechnology.common.forms
