+++ touchtechnology/common/tests/test_forms.py
@@ -10,13 +10,12 @@ from django.core.exceptions import ValidationError
 from django.test import TestCase
 from django.test.utils import override_settings
 from django.utils import timezone
-from touchtechnology.common.forms import (
-    EmailField, iCheckSelectMultiple, permissionformset_factory,
-)
+from touchtechnology.common.forms.auth import permissionformset_factory
 from touchtechnology.common.forms.fields import (
-    BooleanChoiceField, HTMLField, SelectDateField, SelectDateTimeField,
-    SelectTimeField,
+    BooleanChoiceField, EmailField, HTMLField, SelectDateField,
+    SelectDateTimeField, SelectTimeField,
 )
+from touchtechnology.common.forms.icheck import iCheckSelectMultiple
 from touchtechnology.common.models import SitemapNode
 from touchtechnology.common.tests import factories
