+++ touchtechnology/common/tests/test_utils.py
@@ -5,9 +5,8 @@ import mock
 import six
 from test_plus import TestCase
 from touchtechnology.common.forms.fields import boolean_choice_field_coerce
-from touchtechnology.common.utils import (
-    average, get_base_url, get_mod_func, timezone_choice,
-)
+from touchtechnology.common.forms.tz import timezone_choice
+from touchtechnology.common.utils import average, get_base_url, get_mod_func
 
 
 class Average(TestCase):
